[IMP] hr_employee_lastnames: rename hooks file
Rename file of hooks to follow guidelines.
