+++ hr_employee_lastnames/__init__.py
@@ -1,2 +1,2 @@
 from . import models
-from .hook import post_init_hook
+from .hooks import post_init_hook
