+++ hr_employee_lastnames/__manifest__.py
@@ -3,7 +3,7 @@
     "name": "HR Employee First Name and Two Last Names",
     "version": "16.0.1.0.1",
     "author": "Vauxoo, Odoo Community Association (OCA)",
-    "maintainer": "Vauxoo",
+    "maintainers": ["luisg123v"],
     "website": "https://github.com/OCA/hr",
     "license": "AGPL-3",
     "category": "Human Resources",
